requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.4.0
billiard==4.2.3
celery==5.5.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.3
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.10.1
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
numpy==2.3.4
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
PyYAML==6.0.3
qrcode==8.2
redis==4.6.0
referencing==0.37.0
requests==2.32.3
rpds-py==0.29.0
six==1.17.0
sqlparse==0.5.3
timedelta==2020.12.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.14
